[update] float variable.

+++ src/main.cpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 int main(void) {
-    int x = 5.2;
+    float x = 5.2;
 
     std::cout << x << std::endl;
